requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Cloud Providers
boto3==1.34.0
azure-mgmt-resource==23.0.1
google-cloud-storage==2.14.0

# Security Scanning
trivy==0.48.1
bandit==1.7.6
safety==2.3.5

# Threat Intelligence
requests==2.31.0
vt-py==0.17.4
otxv2==1.5.12

# Container Security
docker==7.0.0

# Network Analysis
scapy==2.5.0
pyshark==0.6
python-nmap==0.7.1

# ML/AI
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
flake8==7.0.0
black==24.1.1
mypy==1.8.0
